+++ node/service/src/chain_spec/crab.rs
@@ -187,7 +187,7 @@ pub fn genesis_config() -> ChainSpec {
 			ethereum: Default::default(),
 			base_fee: Default::default(),
 			from_darwinia_issuing: Default::default(),
-			to_parachain_backing: Default::default(),
+			to_crab_parachain_backing: Default::default(),
 		}
 	}
 
@@ -282,7 +282,7 @@ pub fn development_config() -> ChainSpec {
 			ethereum: Default::default(),
 			base_fee: Default::default(),
 			from_darwinia_issuing: Default::default(),
-			to_parachain_backing: Default::default(),
+			to_crab_parachain_backing: Default::default(),
 		}
 	}
