+++ node/service/src/service/mod.rs
@@ -311,7 +311,6 @@ where
 			prometheus_registry.as_ref(),
 			telemetry.as_ref().map(|x| x.handle()),
 		);
-		let client_clone = client.clone();
 		let slot_duration = babe_link.config().slot_duration();
 		let babe_config = BabeParams {
 			keystore: keystore_container.sync_keystore(),
@@ -321,19 +320,15 @@ where
 			env: proposer,
 			sync_oracle: network.clone(),
 			justification_sync_link: network.clone(),
-			create_inherent_data_providers: move |parent, ()| {
-				let client_clone = client_clone.clone();
+			create_inherent_data_providers: move |_, ()| async move {
+				let timestamp = sp_timestamp::InherentDataProvider::from_system_time();
+				let slot =
+					sp_consensus_babe::inherents::InherentDataProvider::from_timestamp_and_duration(
+						*timestamp,
+						slot_duration,
+					);
 
-				async move {
-					let timestamp = sp_timestamp::InherentDataProvider::from_system_time();
-					let slot =
-						sp_consensus_babe::inherents::InherentDataProvider::from_timestamp_and_duration(
-							*timestamp,
-							slot_duration,
-						);
-
-					Ok((timestamp, slot))
-				}
+				Ok((timestamp, slot))
 			},
 			force_authoring,
 			backoff_authoring_blocks,
